move transaction ref to a separate file
