+++ crates/starfish/core/src/error.rs
@@ -9,8 +9,9 @@ use thiserror::Error;
 use typed_store::TypedStoreError;
 
 use crate::{
-    block_header::{BlockRef, GENESIS_ROUND, GenericTransactionRef, Round},
+    block_header::{BlockRef, GENESIS_ROUND, Round},
     commit::{Commit, CommitIndex},
+    transaction_ref::GenericTransactionRef,
 };
 
 /// Errors that can occur when processing blocks, reading from storage, or
