+++ crates/starfish/core/src/lib.rs
@@ -37,6 +37,7 @@ mod threshold_clock;
 mod transaction;
 #[cfg(msim)]
 pub mod transaction;
+mod transaction_ref;
 mod transactions_synchronizer;
 
 mod universal_committer;
